requirements.txt
Babel==2.3.4
coverage==4.2
Django==1.10.4
django-autocomplete-light==3.1.6
django-ckeditor==5.0.3
django-modeltranslation==0.12
django-mptt==0.8.6
django-phonenumber-field==1.1.0
django-recaptcha==1.0.5
django-rosetta==0.7.12
django-webpack-loader==0.3.3
djangorestframework==3.4.6
factory-boy==2.6.1
fake-factory==0.5.10
flake8==3.0.3
ipaddress==1.0.16
lorem-ipsum-generator==0.3
lxml==3.6.4
MarkupSafe==0.23
mccabe==0.5.1
microsofttranslator==0.8
nodeenv==0.13.6
pep8==1.7.0
phonenumbers==7.5.1
phonenumberslite==7.5.1
Pillow==3.3.0
polib==1.0.7
psycopg2==2.6.2
pycodestyle==2.0.0
pyflakes==1.2.3
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.10.0
six==1.10.0
transliterate==1.8.1
uWSGI==2.0.13.1
xlrd==1.0.0
